backend/requirements.txt
fastapi
uvicorn
scrapling
groq
python-dotenv
pandas
pydantic
python-multipart
aiohttp
# removed playwright manual install since user asked not to use it
